move ICompiler and IBlock to plugins.ts
