+++ src/index.ts
@@ -6,6 +6,5 @@
  */
 
 export * from "./client";
-export * from "./compiler";
 export * from "./documentalist";
 export * from "./plugins";
